requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core
azure-identity==1.2.0
azure-functions
azure-storage-file-datalake
numpy==1.19.3
pandas
requests